backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator==2.3.0
google-generativeai>=0.8.0
geopy>=2.4.0

